refactor: extract font-readiness into core controller
First slice of moving orchestration out of the 6k-line PagedEditor.tsx God component into the framework-neutral core controller. Move the initial-layout font-readiness logic (font-face collection from the document model + PM content, the load gate, and the browser font-set probe) into @stll/folio-core/controller/fontReadiness. The existing font-collection test moves to core alongside the code (plus two new cases: the SSR/headless guard and the null-model default floor). PagedEditor imports the functions instead of defining them; INITIAL_FONT_READY_SUPPRESSION_MS (React effect policy) stays behind. Arch tests (react-free-core, layer-boundaries, model-purity) green; no public API change.
